refactor to useDB
