+++ api/schemas/degree.ts
@@ -1,4 +1,4 @@
-import { Schema } from 'mongoose';
+import { Schema, connection } from 'mongoose';
 
 export interface Degree {
   name: string;
@@ -13,3 +13,6 @@ export const DegreeSchema = new Schema<Degree>({
   total: { type: Number, required: true },
   options: { type: Array, required: true },
 });
+
+const degreeDB = connection.useDb('degreeDB');
+export const DegreeModel = degreeDB.model<Degree>('degree', DegreeSchema);
